Quick fix for failing SQL
A query in tag module fails with error 1055. Discussion here:
https://xoops.org/modules/newbb/viewtopic.php?post_id=364611

This changes the SQL_MODE for the session to allow the offending
ONLY_FULL_GROUP_BY violation.

I needed the tag module to work for theme testing, and this was
the quickest fix I could find. The issue deserves a complete fix,
that may include changes to core and module code.

If more proper changes exist, please ignore this.